[BUGFIX] Conform createItem to the parent interface
Duplicates should not throw an exception, since that exception is
not documented in the interface and 3rd party modules would not
expect it to be thrown. This is specially harmful for modules
deciding the queue class at runtime, since they may not have a
custom integration minding the exception.

Since this is an API change, I will bump version to 2.x.