Initial port of module to Drupal 8 (#5)
Replace existing Drupal 7 files with the corresponding Drupal 8
versions.

Main SystemSetQueue class is replaced with a corresponding
UniqueDatabaseQueue class.

The new class uses the same approach as the Drupal 7 version by 
extending the existing schema with a new md5 field with an unique 
key and interpreting exceptions to be an indication of an existing
item.

The MD5 checksum is calculated using PHP because it is simpler 
(and potentially more portable) when using the Drupal 8 database API.

The implementation is reduced a lot simpler by inheriting more of the 
methods from the parent DatabaseQueue class.

Dependency on registry_autoload is dropped as the module is obsolete
for Drupal 8.

Update README accordingly. References to Queue UI have been dropped
for now as there are no patches for the Drupal 8 version which
can add support for this module.